+++ tests/testthat/test-rc_input_rmf.R
@@ -550,3 +550,27 @@ test_that("rc_input_rmf validates W_ET_REFACT at boundary values", {
                  dt.time = dt_time)
   )
 })
+
+test_that("rc_input_rmf correctly throws error when columns do not have act or ref ET", {
+  # set crop table
+  dt_crop <- data.table(
+    B_LU_START = "2022-04-01",
+    B_LU_END = "2022-09-01"
+  )
+  
+  # set time table
+  dt_time <- rc_time_period(start_date = "2022-01-01", end_date = "2022-12-31")
+  
+  # read in weather table
+  weather <- create_weather()
+  weather <- rc_update_weather(dt = weather, dt.time = dt_time)
+  
+  
+  # Run with no ET_REF_MONTH column and no usable ET_ACT
+  weather_noref <- copy(weather)[, W_ET_REF_MONTH := NULL]
+  if ("W_ET_ACT_MONTH" %in% names(weather_noref)) weather_noref[, W_ET_ACT_MONTH := NA_real_]
+  expect_error(rc_input_rmf(dt = dt_crop, A_CLAY_MI = 20, dt.weather = weather_noref, dt.time = dt_time),
+               "W_ET_REF_MONTH is not available")
+  
+
+})